Config : Move config files to parent folder PKG
